backend/requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
Deprecated==1.2.13
Django==4.0.3
django-cors-headers==3.11.0
django-redis==5.2.0
djangorestframework==3.13.1
mysqlclient==2.1.0
packaging==21.3
pyparsing==3.0.7
pytz==2021.3
redis==4.1.4
sqlparse==0.4.2
tzdata==2021.5
wrapt==1.14.0
